requirements.txt
# AWS & AI
boto3==1.34.0
anthropic==0.18.1

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Visualization
plotly==5.17.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
faker==19.6.2
requests==2.31.0

# Email (for sending emails)
sendgrid==6.11.0  # or use boto3 SES